requirements.txt
fastapi>=0.78.0,<0.79.0
pydantic>=1.9.0,<2.0.0
uvicorn>=0.17.0,<0.18.0
envparse==0.2.0
mysql-connector-python==8.0.26
PyMySQL==1.0.2
mysql-connector==2.2.9
loguru==0.6.0